mtd: spi-nor: spansion: comment the name on s28hs256t's entry
Commit 1dc6cd4f94ad5 ("mtd: spi-nor: spansion: Add support for
S28HS256T") added support for the flash. We no longer set
flash_info.name on new entries. Still, it is useful to know which flash
an entry belongs to, and it is tedious to look up a part name from its
ID. So comment the name on top of the entry.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>