mtd: spi-nor: atmel: comment the name on at25sf321(b)'s entry
792caa4a4baaa ("mtd: spi-nor: atmel: add at25sf321 entry") added support
for the flash. We no longer set flash_info.name on new entries. Still,
it is useful to know which flash an entry belongs to, and it is tedious
to look up a part name from its ID. On top of that, we got some useful
information from the patch that at25sf321 and at25s321b share the same
ID but the former does not support SFDP. Comment both pieces of
information on top of the entry.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>