hugetlb: clarify comment in prep_and_add_bootmem_folios()
"Subdivide locks" is a vague statement. It can be understood to mean
that multiple locks are used, each with its own portion to control. This
is confusing and unclear on its own.

The comment was added in b78b27d02930f ("hugetlb: parallelize 1G hugetlb
initialization"), and in the context of that commit it does make sense.
The commit changed the locking from a single cycle to one cycle per
folio. With the extra context, the "subdivide" does make sense.

Improve the comment to stand on its own without a need for digging
through git history.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>