hugetlb: make hugetlb_file_setup() use kernel style error handling
It is common for a lot of kernel code to call a function, check its
value, and if it is an error, return out or goto an error handling
block. Then after that error check, the code can proceed assuming the
good state. Update hugetlb_file_setup() to do the same. This makes it
easier to grok for someone already used to this style. No functional
change indended.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>