luo: memfd: use vmalloc for tracking preserved folios
Currently the preserved folios array is allocated from the FDT. The FDT
is allocated and preserved as a folio. So the size of the FDT is capped
by the maximum folio size. With the default max order of 10 and 4k
pages, this makes the max FDT size 4 MiB. Each folio takes up 16 bytes,
so this gives us a rough maximum of 262144 folios, or roughly 1 GiB of
0-order folios. This places an artificial limit on the size of memfd
that can be preserved.

Use vmalloc to get the buffer to hold the folios instead, and preserve
it using kho_preserve_vmalloc(). This doesn't have the size limit and
this removes the 1 GiB size restriction.

Since the FDT now only contains a small amount of metadata, make it
PAGE_SIZE. The number of folios can no longer be determined by FDT
property size, so add a new field to the FDT to store it.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>