luo_file: add private argument to store runtime state
Currently file handlers only get the data field to store their state.
This field has a pointer to the serialized state of the file, and it
becomes a part of LUO file's serialized state.

File handlers can also need some runtime state to track information that
shouldn't make it in the seriaized data.

One such example is a vmalloc pointer. While kho_preserve_vmalloc()
preserves the memory backing a vmalloc allocation, it does not store the
original vmap pointer, since that has no use being passed to the next
kernel. The pointer is needed to free the memory in case the file is
unpreserved.

Provide a private field in struct luo_file and pass it to all the
callbacks. The field's memory is managed by LUO to keep file handler
code simpler. The size of the private data can be specified when
defining the file handler, in its struct liveupdate_file_ops.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>