luo_file: rename private_data to data and clarify its usage
The private_data field in a luo_file is not used by the core to store
any state between phases despite what its documentation says. The core
treats the field completely opaque and only ever lets callers fill it
in. The value that the callers fill in is then saved in the serialized
LUO file descriptor.

The filed actually holds a handle to the serialized state of the file.
For example, for memfd it contains the physical address of the FDT
containing the serialized state.

Update the documentation to reflect that. Also rename it to data to
match up with the equivalent field in struct luo_file_ser. Both names
are equally as vague, so might as well make them consistent.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>