luo_file: rename private_data to serialized_data and clarify its usage
The private_data field in a luo_file is used to store the u64 handle to
the file's serialized state. The value that the callers fill in is then
saved in the serialized LUO file descriptor.

private_data or similar names are often used to store extra runtime
state for the caller. A follow up patch will add something similar. This
rename makes the roles of both fields clear.

Update the field name in liveupdate_file_op_args as well for
consistency.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>