luo_file: do not reset data on cancel
When LUO is cancelled, the cancel callback is fired before unpreserve,
but unpreserve needs to use data. In fact, unpreserve is the callback
that should free the memory so it should only be reset after that.

Since unpreserve is immediately followed by freeing the file, there is
no need to reset the data field at all and can be left as-is.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>