luo_file: drop file from session on retrieve
When a file is retrieved, its state is deserialized and the preserved
metadata is unpreserved and freed. If the file is left on the session's
file list, the file will get a unpreserve callback when the session is
closed.

The unpreserve callback is used to undo the work done by preserve. This
can include freeing some metadata. For example, among other things,
memfd unpreserve unpins the folios. On a post-kexec memfd, the folios
are never pinned. Calling unpreserve on this memfd would cause the folio
pin count to underflow.

Retrieve is a inherently different operation. Treat retrieve as a
unpreserve as well and remove the file from the session list.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>