mm: memfd_luo: clean up flag handling and move them to ABI header
Currently, the memfd preservation code uses a set of macros to pack PFN
and flags into one u64. Use bitfields instead. They are easier to use
and make the split clearer.

The flags are also part of the ABI. Move them to the ABI header. Rename
them to make their purpose clearer.

Signed-off-by: Pratyush Yadav <pratyush@kernel.org>